requirements.txt
Django~=3.2.5
djangorestframework~=3.12.4
requests~=2.26.0
APScheduler~=3.7.0